add doc for `autd3::controller`
