+++ autd3/src/lib.rs
@@ -1,3 +1,8 @@
+#![warn(missing_docs)]
+#![warn(rustdoc::missing_crate_level_docs)]
+#![warn(rustdoc::unescaped_backticks)]
+
+/// [`Controller`] module.
 pub mod controller;
 pub mod datagram;
 pub mod error;
